fix: refactor strain_design.core to cameo.core.strain_design and cameo.core.target
